fix warnings (#937)

+++ include/wex/factory/stc.h
@@ -2,7 +2,7 @@
 // Name:      stc.h
 // Purpose:   Declaration of class wex::factory::stc
 // Author:    Anton van Wezenbeek
-// Copyright: (c) 2020-2024 Anton van Wezenbeek
+// Copyright: (c) 2020-2025 Anton van Wezenbeek
 ////////////////////////////////////////////////////////////////////////////////
 
 #pragma once
@@ -342,7 +342,7 @@ protected:
   int  m_margin_text_click{-1};
   bool m_margin_text_is_shown{false};
 
-  const size_t m_margin_divider_number{1}, m_margin_folding_number{2},
+  const int m_margin_divider_number{1}, m_margin_folding_number{2},
     m_margin_line_number{0}, m_margin_text_number{3};
 
   int m_saved_pos{-1}, m_saved_selection_start{-1}, m_saved_selection_end{-1};
